refactor: Merge `ProxyValidate` and `ServiceValidate`.
