+++ src/Handler/ServiceValidate.php
@@ -11,31 +11,69 @@ declare(strict_types=1);
 
 namespace EcPhp\CasLib\Handler;
 
-use EcPhp\CasLib\Utils\Uri;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\UriInterface;
+use EcPhp\CasLib\Contract\Response\Type\ServiceValidate as TypeServiceValidate;
+use EcPhp\CasLib\Exception\CasException;
+use EcPhp\CasLib\Exception\CasHandlerException;
+use Exception;
+use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Throwable;
 
-final class ServiceValidate extends Service implements ServiceValidateHandlerInterface
+final class ServiceValidate extends Handler implements ServiceValidateHandlerInterface
 {
-    protected function getProtocolProperties(UriInterface $uri): array
+    public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        $protocolProperties = $this->getProperties()['protocol']['serviceValidate'] ?? [];
+        try {
+            $response = $this->getClient()->sendRequest($request);
+        } catch (ClientExceptionInterface $exception) {
+            throw CasException::errorWhileDoingRequest($exception);
+        }
 
-        $protocolProperties['default_parameters'] += [
-            'service' => (string) $uri,
-            'ticket' => Uri::getParam($uri, 'ticket'),
-        ];
+        $response = $this->getCasResponseBuilder()->fromResponse($response);
 
-        return $protocolProperties;
-    }
+        if (false === ($response instanceof TypeServiceValidate)) {
+            throw CasHandlerException::serviceValidateValidationFailed();
+        }
 
-    protected function getUri(RequestInterface $request): UriInterface
-    {
-        return $this
-            ->buildUri(
-                $request->getUri(),
-                'serviceValidate',
-                $this->formatProtocolParameters($this->getParameters($request))
+        try {
+            /** @var TypeServiceValidate $response */
+            $proxyGrantingTicket = $response->getProxyGrantingTicket();
+        } catch (Throwable $exception) {
+            return $response;
+        }
+
+        $hasPgtIou = $this
+            ->getCache()
+            ->hasItem($proxyGrantingTicket);
+
+        if (false === $hasPgtIou) {
+            throw new Exception('PGT not found in the cache.');
+        }
+
+        try {
+            $pgtId = $this
+                ->getCache()
+                ->getItem($proxyGrantingTicket);
+        } catch (Throwable $exception) {
+            throw new Exception('Unable to get PGT from cache', 0, $exception);
+        }
+
+        if (null === $pgtId->get()) {
+            throw new Exception('Invalid PGT ID value.');
+        }
+
+        return $response
+            ->withBody(
+                $this
+                    ->getPsr17()
+                    ->createStream(
+                        str_replace(
+                            $proxyGrantingTicket,
+                            $pgtId->get(),
+                            (string) $response->getBody()
+                        )
+                    )
             );
     }
 }
